Adding support for connection pooling in TPCC. (#25)
* Adding support for connection pooling in TPCC.

Summary:
We need 10X the number of terminals with respect to the number of
warehouses to get a good TPM-C number. Each terminal gets its own
connection to perform transactions. This is fine for smaller number of
warehouses. But as the number increases, we reach the limit of the max
number of connections supported by the DB.

Given that each terminal spends most of its time waiting for keying or
for thinking, we could reuse the connections across the terminals.

This change creates a Hikari pool one per endpoint and each worker gets
a reference to the pool rather than an actual connection. When the
worker executes a transaction, it gets a connection from the pool and
 gives the connection back to the pool once the transaction is done.

Each worker prepares a statement every time it executes a transaction.
This is fine since the driver has a cache of PreparedStatements per
connection of size 256. The number of unique statements in the
application is 33. Hence we don't send the PrepareStatement requests to
the server every time.

Reviewers:
Neha, Mikhail, Karthik