Making configuration easier for the benchmark.
Summary:
Currently for running the benchmark we will have to configure a lot of
parameters in the config file and then run the benchmark.

With this change we can run the workload by specifying the list of nodes
and the number of warehouses.
example load:
./tpccbenchmark --create=true --load=true
./tpccbenchmark --create=true --load=true --nodes=127.0.0.1,127.0.0.2
./tpccbenchmark --create=true --load=true --nodes=127.0.0.1,127.0.0.2 --warehouses=20 --loaderthreads=20

example execute:
./tpccbenchmark --execute=true
./tpccbenchmark --execute=true --nodes=127.0.0.1,127.0.0.2
./tpccbenchmark --execute=true --nodes=127.0.0.1,127.0.0.2 --warehouses=20

The default values for
nodes -> 127.0.0.1
warehouses -> 10
loaderthreads -> 10
terminals -> 10 * warehouses
DBconnections -> warehouses

The values for terminals and DBconnections can be further tweaked in the
config file.

Reviewers:
Neha, Karthik