Reduce the memory consumption of TPC-C benchmark.
Summary:
The original code was pre allocating a large array for the latency
results assuming that they were not going to use a lot of threads.
However, in the TPCC case we have a large number with each thread doing
very small number of operations. So this change reduces the size of the
arrays.

Removed some other components like histograms that were not being used
that contribute to the memory bloat.

Added some more cleanup of the code as well.

Reviewers:
Neha