Flush updates to Stock if the same item is present multiple times in the batch of items. (#53)
Summary:
Item ids are generated randomly and it is possible that the same item
can be present multiple times in an order. In that case the updates to
the Stock table doesn't reflect the correct values, since we batch the
updates and not send the request until the end. So the second update
ends up reading the stale value of the Stock table row.

Fix this by flushing the outstanding batches if we encounter the same
item again in the batch.

Reviewers:
Mihnea