Use SQL procedures for updates to the STOCK table.
Summary:
Currently we batch the updates only when they are part of a SQL
procedure. Given that we have to perform 10 updates on an average in a
NewOrder transaction we use the stored procedures to perform these
updates.

The stored procedures are created after the Loading of all the tables
are done.

The usage of these stored procedures is controlled by the
'useStoredProcedures' argument in the xml config

Reviewers:
Mihnea