Multiple prepare statements issue during inserts (#122)
Number prepared statements for every insert statement should be equal to number of retries. If there is an exception in insert, same prepared statement cannot be used to retry in certain cases. So for every retry we will use a different prepared statement. Clear all the prepared statement after one is successful.

Not the best way to solve the problem but lot of re-organizing would be needed if we want a different approach.