Remove call to getTransactionIsolation as it is not needed and is throwing an error #138
There is a call for conn.getTransactionIsolation() from tpcc code. This code gets executed only when debug logging is enabled. This call to conn.getTransactionIsolation() is throwing an exception and looks like a driver issue. This code is unimportant in tpcc and hence fixing the issue for now by removing the code.